Add changes according to the review
