+++ src/plugins/intel_npu/src/al/include/intel_npu/icompiler.hpp
@@ -148,32 +148,6 @@ struct NetworkDescription final {
 };
 
 /**
- * @struct CompiledNetwork
- * @brief Custom container for compiled network, used for export
- * @var CompiledNetwork::data
- * Pointer to the address of compiled network
- * @var CompiledNetwork:size
- * Size of the compiled network
- * @var CompiledNetwork::ownedStorage
- * Plugin owned compiled network storage that is required in case of a driver that
- * doesn't support graph extension 1.7, as in this case plugin must create a copy of the compiled network.
- * @note It's unsafe to store either data or size outside of the compiled network object as its destructor
- * would release the owning container
- */
-
-struct CompiledNetwork {
-    const uint8_t* data;
-    size_t size;
-    CompiledNetwork(const uint8_t* data, size_t size, std::vector<uint8_t> storage)
-        : data(data),
-          size(size),
-          ownedStorage(std::move(storage)) {}
-
-private:
-    std::vector<uint8_t> ownedStorage;
-};
-
-/**
  * @interface ICompiler
  * @brief An interface to be implemented by a concrete compiler to provide
  * methods for preparing a network for execution on a NPU device
