+++ src/plugins/intel_npu/src/backend/include/zero_device.hpp
@@ -7,7 +7,6 @@
 #include <ze_api.h>
 #include <ze_graph_ext.h>
 
-#include "intel_npu/common/iadapter.hpp"
 #include "intel_npu/common/icompiled_model.hpp"
 #include "intel_npu/common/npu.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
@@ -33,10 +32,7 @@ public:
     std::map<ov::element::Type, float> getGops() const override;
     ov::device::Type getDeviceType() const override;
 
-    std::shared_ptr<IAdapter> createAdapter() override;
-
     std::shared_ptr<SyncInferRequest> createInferRequest(const std::shared_ptr<const ICompiledModel>& compiledModel,
-                                                         const std::shared_ptr<IGraph>& graph,
                                                          const Config& config) override;
     void updateInfo(const Config& config) override {
         log.setLevel(config.get<LOG_LEVEL>());
