+++ src/plugins/intel_npu/src/backend/src/zero_device.cpp
@@ -7,7 +7,6 @@
 #include "intel_npu/common/itt.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
 #include "intel_npu/utils/zero/zero_utils.hpp"
-#include "zero_adapter.hpp"
 #include "zero_host_tensor.hpp"
 #include "zero_infer_request.hpp"
 #include "zero_remote_tensor.hpp"
@@ -171,37 +170,10 @@ ov::device::Type ZeroDevice::getDeviceType() const {
     return ov::device::Type::INTEGRATED;
 }
 
-std::shared_ptr<IAdapter> ZeroDevice::createAdapter() {
-    switch (_initStructs->getGraphDdiTable().version()) {
-    case ZE_GRAPH_EXT_VERSION_1_3:
-        return std::make_shared<ZeroAdapter<ze_graph_dditable_ext_1_3_t>>(_initStructs);
-        break;
-    case ZE_GRAPH_EXT_VERSION_1_4:
-        return std::make_shared<ZeroAdapter<ze_graph_dditable_ext_1_4_t>>(_initStructs);
-        break;
-    case ZE_GRAPH_EXT_VERSION_1_5:
-        return std::make_shared<ZeroAdapter<ze_graph_dditable_ext_1_5_t>>(_initStructs);
-        break;
-    case ZE_GRAPH_EXT_VERSION_1_6:
-        return std::make_shared<ZeroAdapter<ze_graph_dditable_ext_1_6_t>>(_initStructs);
-        break;
-    case ZE_GRAPH_EXT_VERSION_1_7:
-        return std::make_shared<ZeroAdapter<ze_graph_dditable_ext_1_7_t>>(_initStructs);
-        break;
-    case ZE_GRAPH_EXT_VERSION_1_8:
-        return std::make_shared<ZeroAdapter<ze_graph_dditable_ext_1_8_t>>(_initStructs);
-        break;
-    default:
-        return std::make_shared<ZeroAdapter<ze_graph_dditable_ext_1_2_t>>(_initStructs);
-        break;
-    }
-}
-
 std::shared_ptr<SyncInferRequest> ZeroDevice::createInferRequest(
     const std::shared_ptr<const ICompiledModel>& compiledModel,
-    const std::shared_ptr<IGraph>& graph,
     const Config& config) {
-    return std::make_shared<ZeroInferRequest>(_initStructs, compiledModel, graph, config);
+    return std::make_shared<ZeroInferRequest>(_initStructs, compiledModel, config);
 }
 
 ov::SoPtr<ov::IRemoteTensor> ZeroDevice::createRemoteTensor(std::shared_ptr<ov::IRemoteContext> context,
