+++ src/plugins/intel_npu/src/common/include/intel_npu/common/npu.hpp
@@ -6,12 +6,10 @@
 
 #include <cstdint>
 
-#include "intel_npu/common/iadapter.hpp"
 #include "intel_npu/common/icompiled_model.hpp"
 #include "intel_npu/common/igraph.hpp"
 #include "intel_npu/common/sync_infer_request.hpp"
 #include "intel_npu/config/config.hpp"
-#include "intel_npu/icompiler.hpp"
 #include "openvino/runtime/intel_npu/remote_properties.hpp"
 #include "openvino/runtime/iremote_context.hpp"
 #include "openvino/runtime/properties.hpp"
@@ -84,11 +82,8 @@ public:
     virtual ov::device::Type getDeviceType() const;
     virtual std::map<ov::element::Type, float> getGops() const;
 
-    virtual std::shared_ptr<IAdapter> createAdapter();
-
     virtual std::shared_ptr<SyncInferRequest> createInferRequest(
         const std::shared_ptr<const ICompiledModel>& compiledModel,
-        const std::shared_ptr<IGraph>& graph,
         const Config& config) = 0;
 
     virtual void updateInfo(const Config& config) = 0;
