+++ src/plugins/intel_npu/src/compiler_adapter/include/driver_compiler_adapter.hpp
@@ -6,15 +6,22 @@
 
 #pragma once
 
-#include "intel_npu/common/iadapter.hpp"
+#include <ze_api.h>
+#include <ze_graph_ext.h>
+
+#include <type_traits>
+#include <utility>
+
 #include "intel_npu/common/npu.hpp"
+#include "intel_npu/config/config.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
+#include "izero_adapter.hpp"
 
 namespace intel_npu {
 
 class DriverCompilerAdapter final : public ICompilerAdapter {
 public:
-    DriverCompilerAdapter(const std::shared_ptr<IDevice>& device);
+    DriverCompilerAdapter(const std::shared_ptr<IEngineBackend>& iEngineBackend);
 
     std::shared_ptr<IGraph> compile(const std::shared_ptr<const ov::Model>& model, const Config& config) const override;
 
@@ -23,7 +30,32 @@ public:
     ov::SupportedOpsMap query(const std::shared_ptr<const ov::Model>& model, const Config& config) const override;
 
 private:
-    std::shared_ptr<IAdapter> _adapter;
+    /**
+     * @brief Serialize input / output information to string format.
+     * @details Format:
+     * --inputs_precisions="0:<input1Precision> [1:<input2Precision>]"
+     * --inputs_layouts="0:<input1Layout> [1:<input2Layout>]"
+     * --outputs_precisions="0:<output1Precision>"
+     * --outputs_layouts="0:<output1Layout>"
+     *
+     * For older compiler versions, the name of the inputs/outputs may be used instead of their indices.
+     *
+     * Since the layout information is no longer an important part of the metadata values when using the 2.0 OV
+     * API, the layout fields shall be filled with default values in order to assure the backward compatibility
+     * with the driver.
+     */
+    std::string serializeIOInfo(const std::shared_ptr<const ov::Model>& model, const bool useIndices) const;
+
+    SerializedIR serializeIR(const std::shared_ptr<const ov::Model>& model,
+                             ze_graph_compiler_version_info_t compilerVersion,
+                             const uint32_t supportedOpsetVersion) const;
+
+    std::string serializeConfig(const Config& config, ze_graph_compiler_version_info_t compilerVersion) const;
+
+    std::shared_ptr<IZeroAdapter> _zeroAdapter;
+
+    ze_device_graph_properties_t _deviceGraphProperties = {};
+
     Logger _logger;
 };
