+++ src/plugins/intel_npu/src/compiler_adapter/include/ir_serializer.hpp
