+++ src/plugins/intel_npu/src/compiler_adapter/include/izero_adapter.hpp
@@ -11,13 +11,14 @@
 
 namespace intel_npu {
 
-class IAdapter {
-public:
-    virtual std::unordered_set<std::string> queryResultFromSupportedLayers(
-        std::pair<size_t, std::shared_ptr<uint8_t>> serializedIR,
-        const std::string& buildFlags) const = 0;
+using SerializedIR = std::pair<size_t, std::shared_ptr<uint8_t>>;
 
-    virtual ze_graph_handle_t getGraphHandle(std::pair<size_t, std::shared_ptr<uint8_t>> serializedIR,
+class IZeroAdapter {
+public:
+    virtual std::unordered_set<std::string> queryResultFromSupportedLayers(SerializedIR serializedIR,
+                                                                           const std::string& buildFlags) const = 0;
+
+    virtual ze_graph_handle_t getGraphHandle(SerializedIR serializedIR,
                                              const std::string& buildFlags,
                                              const uint32_t& flags) const = 0;
 
@@ -31,16 +32,14 @@ public:
 
     virtual void setArgumentValue(ze_graph_handle_t graphHandle, uint32_t argi_, const void* argv) const = 0;
 
-    virtual void graphInitialie(ze_graph_handle_t graphHandle, const Config& config) const = 0;
+    virtual void graphInitialize(ze_graph_handle_t graphHandle, const Config& config) const = 0;
 
     virtual std::tuple<std::vector<ArgumentDescriptor>, std::vector<ArgumentDescriptor>> getIODesc(
         ze_graph_handle_t graphHandle) const = 0;
 
     virtual std::shared_ptr<CommandQueue> crateCommandQueue(const Config& config) const = 0;
 
-    virtual ze_device_graph_properties_t getDeviceGraphProperties() const = 0;
-
-    virtual ~IAdapter() = default;
+    virtual ~IZeroAdapter() = default;
 };
 
 }  // namespace intel_npu
