+++ src/plugins/intel_npu/src/compiler_adapter/src/driver_graph.cpp
@@ -2,25 +2,21 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "cip_graph.hpp"
+#include "driver_graph.hpp"
 
 #include "intel_npu/config/common.hpp"
 #include "intel_npu/config/runtime.hpp"
 
 namespace intel_npu {
 
-CipGraph::CipGraph(const std::shared_ptr<IAdapter>& adapter,
-                   const ov::SoPtr<ICompiler>& compiler,
-                   ze_graph_handle_t graphHandle,
-                   NetworkMetadata metadata,
-                   const std::vector<uint8_t> compiledNetwork,
-                   const Config& config)
+DriverGraph::DriverGraph(const std::shared_ptr<IZeroAdapter>& adapter,
+                         ze_graph_handle_t graphHandle,
+                         NetworkMetadata metadata,
+                         const Config& config)
     : IGraph(graphHandle, std::move(metadata)),
       _adapter(adapter),
-      _compiler(compiler),
-      _compiledNetwork(std::move(compiledNetwork)),
       _config(config),
-      _logger("CipGraph", _config.get<LOG_LEVEL>()) {
+      _logger("DriverGraph", _config.get<LOG_LEVEL>()) {
     if (_config.get<CREATE_EXECUTOR>()) {
         initialize();
     } else {
@@ -28,22 +24,22 @@ CipGraph::CipGraph(const std::shared_ptr<IAdapter>& adapter,
     }
 }
 
-CompiledNetwork CipGraph::export_blob() const {
-    return CompiledNetwork(_compiledNetwork.data(), _compiledNetwork.size(), _compiledNetwork);
+CompiledNetwork DriverGraph::export_blob() const {
+    return _adapter->getCompiledNetwork(_handle);
 }
 
-std::vector<ov::ProfilingInfo> CipGraph::process_profiling_output(const std::vector<uint8_t>& profData) const {
-    return _compiler->process_profiling_output(profData, _compiledNetwork, _config);
+std::vector<ov::ProfilingInfo> DriverGraph::process_profiling_output(const std::vector<uint8_t>& profData) const {
+    OPENVINO_THROW("Profiling post-processing is not supported.");
 }
 
-void CipGraph::set_argument_value(uint32_t argi, const void* argv) const {
+void DriverGraph::set_argument_value(uint32_t argi, const void* argv) const {
     if (_adapter == nullptr) {
         OPENVINO_THROW("Zero compiler adapter wasn't initialized");
     }
     _adapter->setArgumentValue(_handle, argi, argv);
 }
 
-void CipGraph::initialize() {
+void DriverGraph::initialize() {
     if (_adapter) {
         _logger.debug("Graph initialize start");
 
@@ -51,16 +47,16 @@ void CipGraph::initialize() {
         _command_queue = _adapter->crateCommandQueue(_config);
 
         if (_config.has<WORKLOAD_TYPE>()) {
-            setWorkloadType(_config.get<WORKLOAD_TYPE>());
+            set_workload_type(_config.get<WORKLOAD_TYPE>());
         }
 
-        _adapter->graphInitialie(_handle, _config);
+        _adapter->graphInitialize(_handle, _config);
 
         _logger.debug("Graph initialize finish");
     }
 }
 
-CipGraph::~CipGraph() {
+DriverGraph::~DriverGraph() {
     if (_handle != nullptr) {
         auto result = _adapter->release(_handle);
