+++ src/plugins/intel_npu/src/compiler_adapter/src/ir_serializer.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "graph_transformations.hpp"
+#include "ir_serializer.hpp"
 
 #include <cstdint>
 #include <istream>
