+++ src/plugins/intel_npu/src/compiler_adapter/src/plugin_graph.cpp
@@ -2,21 +2,25 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "cid_graph.hpp"
+#include "plugin_graph.hpp"
 
 #include "intel_npu/config/common.hpp"
 #include "intel_npu/config/runtime.hpp"
 
 namespace intel_npu {
 
-CidGraph::CidGraph(const std::shared_ptr<IAdapter>& adapter,
-                   ze_graph_handle_t graphHandle,
-                   NetworkMetadata metadata,
-                   const Config& config)
+PluginGraph::PluginGraph(const std::shared_ptr<IZeroAdapter>& adapter,
+                         const ov::SoPtr<ICompiler>& compiler,
+                         ze_graph_handle_t graphHandle,
+                         NetworkMetadata metadata,
+                         const std::vector<uint8_t> compiledNetwork,
+                         const Config& config)
     : IGraph(graphHandle, std::move(metadata)),
       _adapter(adapter),
+      _compiler(compiler),
+      _compiledNetwork(std::move(compiledNetwork)),
       _config(config),
-      _logger("CidGraph", _config.get<LOG_LEVEL>()) {
+      _logger("PluginGraph", _config.get<LOG_LEVEL>()) {
     if (_config.get<CREATE_EXECUTOR>()) {
         initialize();
     } else {
@@ -24,22 +28,22 @@ CidGraph::CidGraph(const std::shared_ptr<IAdapter>& adapter,
     }
 }
 
-CompiledNetwork CidGraph::export_blob() const {
-    return _adapter->getCompiledNetwork(_handle);
+CompiledNetwork PluginGraph::export_blob() const {
+    return CompiledNetwork(_compiledNetwork.data(), _compiledNetwork.size(), _compiledNetwork);
 }
 
-std::vector<ov::ProfilingInfo> CidGraph::process_profiling_output(const std::vector<uint8_t>& profData) const {
-    OPENVINO_THROW("Profiling post-processing is not supported.");
+std::vector<ov::ProfilingInfo> PluginGraph::process_profiling_output(const std::vector<uint8_t>& profData) const {
+    return _compiler->process_profiling_output(profData, _compiledNetwork, _config);
 }
 
-void CidGraph::set_argument_value(uint32_t argi, const void* argv) const {
+void PluginGraph::set_argument_value(uint32_t argi, const void* argv) const {
     if (_adapter == nullptr) {
         OPENVINO_THROW("Zero compiler adapter wasn't initialized");
     }
     _adapter->setArgumentValue(_handle, argi, argv);
 }
 
-void CidGraph::initialize() {
+void PluginGraph::initialize() {
     if (_adapter) {
         _logger.debug("Graph initialize start");
 
@@ -47,16 +51,16 @@ void CidGraph::initialize() {
         _command_queue = _adapter->crateCommandQueue(_config);
 
         if (_config.has<WORKLOAD_TYPE>()) {
-            setWorkloadType(_config.get<WORKLOAD_TYPE>());
+            set_workload_type(_config.get<WORKLOAD_TYPE>());
         }
 
-        _adapter->graphInitialie(_handle, _config);
+        _adapter->graphInitialize(_handle, _config);
 
         _logger.debug("Graph initialize finish");
     }
 }
 
-CidGraph::~CidGraph() {
+PluginGraph::~PluginGraph() {
     if (_handle != nullptr) {
         auto result = _adapter->release(_handle);
