+++ src/plugins/intel_npu/src/plugin/include/plugin.hpp
@@ -10,7 +10,6 @@
 
 #include "backends.hpp"
 #include "intel_npu/config/config.hpp"
-#include "intel_npu/icompiler.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
 #include "metrics.hpp"
 #include "openvino/runtime/iplugin.hpp"
@@ -53,7 +52,7 @@ public:
                                     const ov::AnyMap& properties) const override;
 
 private:
-    std::unique_ptr<ICompilerAdapter> getCompiler(const std::shared_ptr<IDevice>& device, const Config& config) const;
+    std::unique_ptr<ICompilerAdapter> getCompiler(const Config& config) const;
 
     std::shared_ptr<NPUBackends> _backends;
