+++ src/plugins/intel_npu/tests/functional/behavior/npu_driver_compiler_adapter/custom_stream.cpp
@@ -8,9 +8,10 @@
 #include "common/functions.h"
 #include "common/npu_test_env_cfg.hpp"
 #include "common_test_utils/node_builders/constant.hpp"
-#include "graph_transformations.hpp"
 #include "intel_npu/config/common.hpp"
+#include "ir_serializer.hpp"
 #include "openvino/opsets/opset11.hpp"
+
 
 using CompilationParams = std::tuple<std::string,  // Device name
                                      ov::AnyMap    // Config
